[core][event] add a generic C++ event recording interface
This factors out the C++ event recording infrastructure (that is part
of exec_event_recorder.{hpp,cpp}) into a "generic" event recorder
class with C++ features. In a subsequent commit, the current execution
event recorder will be changed to use this generic utility class.

The reason for doing this is that currently, there is only a single
recorder (the execution event ring). In other branches, other recorders
exist (e.g., the EVM opcode tracer) which have their own C++ helper
objects. This moves the common functionality to a shared class.

Another thing that happens in this refactoring is that event ring
ownership is made a separate concept from the event ring recorder, and
is managed by the RAII helper "OwnedEventRing." Previously, the
exclusive recorder was also responsible for owning the event ring file.
Separating the two concepts simplifies the initialization code when
there are multiple rings.